Add function to get all polybar IPC paths
get_polybar_ipc_paths searches a base directory for polybar_mqueue files and
returns an array with full paths to each of the IPC files.

Also added function to join two paths together with basic logic to check if
the first path already has a forward slash at the end.