Add function to update track info on track change
With the addition of the spotify status enum for preventing the spam of
duplicate messages to polybar, when track changes were made, but spotify's
status remained the same i.e. still playing on track change, then no messages
were sent to polybar to update the track info. This separate function solves
the issue of spamming polybar for updating the controls while keeping the module
up-to-date with track changes.