Add str_replace_all utility function
This function will be used to replace tokens such as %artist% and %title% with
the actual artist name and track title.
To keep this function optimized which avoiding allocating too much memory or
allocating too little and having to keep reallocating memory for the replaced
string, initially, only the memory needed to make one replacement is allocated.
In most use cases of this function, only one replacement will need to be made,
so this prevents any additional realloc calls.

A REALLOC_RATE constant is also available to set to set the rate at which
memory is reallocated for the final string when more memory is needed. The rate
is a measure of the number of replacements for which additional memory will be
allocated and also controls the initial amount of memory allocated.