Fix str_trunc related memory leak
Since str_trunc reallocates memory to reduce the size of the string, the caller
may not have the correct address resulting in other problems. Now str_trunc
returns the address to the new truncated string rather than a boolean. If the
trunc string is longer than max_len, the function returns a NULL pointer, so an
additional check is now set in the caller function.