Fix another memory leak in str_replace_all
Due to the way memory is allocated in advance assuming one replacement, if there
are no replacements and one replacement would result in less memory being
required (REPL_DIFF < 0), there would not be enough memory allocated to copy
over substr to new_str.

Reordering the commands, so that the reallocation takes place before the
concatanation fixes this issue since it would reallocate the exact memory
required before substr is concatenated to new_str.