Add num_of_matches utility function
This function simply returns the number of times a particular string is found
in another string using the same algorithm is the str_replace_all function.

This function will not find matches that are part of other matches. For example,
the string "strstrstr" would be reported as only having 1 match.