Don't truncate if output fits under max_length
If max_length is specified for the output, don't truncate the artist or title
if the untruncated output satisfies the max_length setting. If max_length is
not specified, artist and title will be truncated as expected.

Also set max length for artist, title, and output at MAX_INT by default.

Also update print_usage function with new logic and add some more examples to
clarify.